Extract reusable context snapshot test helpers
